// src/models/tastings.ts
export class Tastings {
    tastingid: number; 
    shopid: number;
    productid:number; 
    weekday:number;  
  /*createdon: any; 
    cratedby:number; 
    modifiedon:any; 
    modifiedby:number; 
    deletedon: any; 
    deletedby: number; */
    isdeleted: boolean; 
    isactive: boolean; 
}
export class Tasting_Increase {
    tastingincreaseid: number; 
    shopid: number;
    productid:number; 
    increment:number;  
 /* createdon: any; 
    cratedby:number; 
    modifiedon:any; 
    modifiedby:number; 
    deletedon: any; 
    deletedby: number; */
    isdeleted: boolean; 
    isactive: boolean; 
}